# M4-Summative-Marta-Polak/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  description: Group 5 App Design M4-Summative-Rodney-Plancher
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /tShirts/{id}:
    get:
      tags:
      - t-shirts-controller
      operationId: getShirtsByID
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TShirts'
    put:
      tags:
      - t-shirts-controller
      operationId: updateShirts
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TShirts'
        required: true
      responses:
        "202":
          description: Accepted
    delete:
      tags:
      - t-shirts-controller
      operationId: deleteShirts
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
  /games/{id}:
    get:
      tags:
      - games-controller
      operationId: getGameByID
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Games'
    put:
      tags:
      - games-controller
      operationId: updateGame
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Games'
        required: true
      responses:
        "202":
          description: Accepted
    delete:
      tags:
      - games-controller
      operationId: deleteGame
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
  /consoles/{id}:
    get:
      tags:
      - consoles-controller
      operationId: getConsoleByID
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Consoles'
    put:
      tags:
      - consoles-controller
      operationId: updateConsoles
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Consoles'
        required: true
      responses:
        "202":
          description: Accepted
    delete:
      tags:
      - consoles-controller
      operationId: deleteConsoles
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
  /tShirts:
    get:
      tags:
      - t-shirts-controller
      operationId: gettinAllTShirts
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TShirts'
    post:
      tags:
      - t-shirts-controller
      operationId: createShirts
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TShirts'
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TShirts'
  /games:
    get:
      tags:
      - games-controller
      operationId: gettingAllGames
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Games'
    post:
      tags:
      - games-controller
      operationId: createGame
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Games'
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Games'
  /consoles:
    get:
      tags:
      - consoles-controller
      operationId: getAllConsoles
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Consoles'
    post:
      tags:
      - consoles-controller
      operationId: createConsoles
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Consoles'
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Consoles'
  /tShirts/size/{size}:
    get:
      tags:
      - t-shirts-controller
      operationId: findBySize
      parameters:
      - name: size
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TShirts'
  /tShirts/color/{color}:
    get:
      tags:
      - t-shirts-controller
      operationId: findByColor
      parameters:
      - name: color
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TShirts'
  /salesTaxRate/{state}:
    get:
      tags:
      - sales-tax-rate-controller
      operationId: getSalesTaxByState
      parameters:
      - name: state
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SalesTaxRate'
  /processingFee/productType:
    get:
      tags:
      - processing-fee-controller
      operationId: getProcessingFee
      parameters:
      - name: productType
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProcessingFee'
  /games/title/{title}:
    get:
      tags:
      - games-controller
      operationId: findGamesByTitle
      parameters:
      - name: title
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Games'
  /games/studio/{studio}:
    get:
      tags:
      - games-controller
      operationId: getGameByStudio
      parameters:
      - name: studio
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Games'
  /games/esrbrating/{esrbrating}:
    get:
      tags:
      - games-controller
      operationId: findGamesByESRBRating
      parameters:
      - name: esrbrating
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Games'
  /consoles/manufacturer/{manufacturer}:
    get:
      tags:
      - consoles-controller
      operationId: findShirtsByManufacturer
      parameters:
      - name: manufacturer
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Consoles'
components:
  schemas:
    TShirts:
      required:
      - color
      - description
      - price
      - quantity
      - size
      type: object
      properties:
        id:
          type: integer
          format: int32
        size:
          type: string
        color:
          type: string
        description:
          type: string
        price:
          type: number
        quantity:
          type: integer
          format: int32
    Games:
      required:
      - description
      - price
      - quantity
      - studio
      - title
      type: object
      properties:
        id:
          type: integer
          format: int32
        title:
          type: string
        description:
          type: string
        price:
          type: number
        studio:
          type: string
        quantity:
          type: integer
          format: int32
        esrbrating:
          type: integer
          format: int32
    Consoles:
      required:
      - manufacturer
      - memoryAmount
      - model
      - price
      - processor
      - quantity
      type: object
      properties:
        id:
          type: integer
          format: int32
        model:
          type: string
        manufacturer:
          type: string
        memoryAmount:
          type: string
        processor:
          type: string
        price:
          type: number
        quantity:
          type: integer
          format: int32
        memory_amount:
          type: string
          writeOnly: true
    SalesTaxRate:
      required:
      - rate
      - state
      type: object
      properties:
        state:
          type: string
        rate:
          type: number
    ProcessingFee:
      required:
      - fee
      - productType
      type: object
      properties:
        productType:
          type: string
        fee:
          type: number